HomeControllerTest moved to appropriate package
