+++ src/test/java/com/glomozda/machinerepair/controller/AllControllerTests.java
@@ -4,7 +4,7 @@ import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
 
-import com.glomozda.machinerepair.controller.adminpage.HomeControllerTest;
+import com.glomozda.machinerepair.controller.HomeControllerTest;
 
 @RunWith(Suite.class)
 @SuiteClasses({ HomeControllerTest.class
